# docs/swagger.yaml
basePath: /
definitions:
  models.CreateProductRequest:
    properties:
      category_id:
        type: integer
      description:
        type: string
      images:
        items:
          type: string
        type: array
      name:
        maxLength: 200
        minLength: 1
        type: string
      sizes:
        items:
          $ref: '#/definitions/models.ProductSize'
        type: array
      stock:
        minimum: 0
        type: integer
      variants:
        items:
          type: integer
        type: array
    required:
    - name
    - stock
    type: object
  models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  models.ProductSize:
    properties:
      price:
        minimum: 0
        type: number
      size_id:
        type: integer
      size_name:
        type: string
    required:
    - price
    - size_id
    type: object
  models.RegisterRequest:
    properties:
      address:
        type: string
      email:
        type: string
      password:
        type: string
      phone:
        type: string
      username:
        type: string
    required:
    - address
    - email
    - phone
    - username
    type: object
  models.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8082
info:
  contact: {}
  description: API documentation for Coffee Shop App
  title: CoffeeShop API
  version: "1.0"
paths:
  /admin/products:
    post:
      consumes:
      - application/json
      description: Admin can create new product
      parameters:
      - description: Product Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Create product
      tags:
      - Products (Admin)
  /admin/products/{id}:
    delete:
      description: Admin can delete product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Delete product
      tags:
      - Products (Admin)
    put:
      consumes:
      - application/json
      description: Admin can update product data
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated Product Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Update product
      tags:
      - Products (Admin)
  /admin/products/{id}/image:
    post:
      consumes:
      - multipart/form-data
      description: Upload image for product (.jpg, .jpeg, .png max 10MB)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Product Image
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Upload product image
      tags:
      - Products (Admin)
  /admin/users:
    get:
      description: Only admin can view all user list
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get all users
      tags:
      - Admin
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login with email and password to get token
      parameters:
      - description: Login Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      summary: Login user
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: Register Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      summary: Register a new user
      tags:
      - Auth
  /products:
    get:
      consumes:
      - application/json
      description: Get products with pagination, search, and sorting
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Limit per page (default: 10)'
        in: query
        name: limit
        type: integer
      - description: Search by product name
        in: query
        name: search
        type: string
      - description: Sort by field
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get product
      tags:
      - Products
  /products/{id}:
    get:
      description: Get detailed info for a single product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get product detail
      tags:
      - Products
  /user/{id}:
    put:
      consumes:
      - application/json
      description: User can update their account, admin can update any account
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Response'
      summary: Update user data
      tags:
      - User
  /user/{id}/picture:
    post:
      consumes:
      - multipart/form-data
      description: Upload profile picture (.jpg/.jpeg/.png)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Profile Picture
        in: formData
        name: picture
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Response'
      summary: Upload user profile picture
      tags:
      - User
swagger: "2.0"